+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+### 0.14.1 (30-Aug-2025)
+
+- Added:
+  - Tests for HardwareBP
+  - Additional tests for BreakAndExec, because of the new single-stepping method:
+    - test\_breakandexec\_classify.py
+    - test\_breakandexec\_filter.py
+
+- Changed:
+  - Signal changed for not enough breakpoints: SIGABRT -> SIGSYS. This means SIGABRT is now only used for Fatal Error.
+  - Signal changed for "executable not loaded": SIGILL -> SIGSEGV. This means that SIGILL is now used exclusively for instructions, from which we cannot continue.
+
 ### 0.14.0 (28-09-2025)
 
 - **Fixed:**